Рефакторит comment-list.js
