+++ js/components/comment-item.js
@@ -1,7 +1,7 @@
 const commentTemplate = document.querySelector('.social__comment').cloneNode(true);
 
 // Создает ноду комментария для последующего рендера
-const createComment = (data) => {
+export const createComment = (data) => {
   const {avatar, name, message} = data;
 
   const newComment = commentTemplate.cloneNode(true);
@@ -13,15 +13,4 @@ const createComment = (data) => {
   textElement.textContent = message;
 
   return newComment;
-};
-// Создает фрагмент с комментариями
-export const createCommentsFragment = (data) => {
-  const commentsFragment = document.createDocumentFragment();
-
-  data.forEach((comment) => {
-    const newComment = createComment(comment);
-    commentsFragment.appendChild(newComment);
-  });
-
-  return commentsFragment;
 };
